% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tom.R
\name{theme_tom}
\alias{theme_tom}
\title{Tom's sweet theme.}
\usage{
theme_tom(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Tom's sweet theme.
}
